requirements.txt
django
djangorestframework
psycopg2-binary
django-filter==22.1
